backend/requirements.txt
# File: backend/requirements.txt
# Version: v0.4.0
# Core scientific/GA stack
biopython>=1.83
# C-accelerated edit distance (preferred)
edlib>=1.3.9.post1
# Optional C++ fallback for Levenshtein distance
rapidfuzz>=3.9.0
# C-backed primer design thermodynamics
primer3-py>=2.0.3
pytest>=8.2.0

# --- API & runtime ---
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7.0
pydantic-settings>=2.3
httpx>=0.27.0
SQLAlchemy>=2.0

ViennaRNA>=2.6.4

numpy>=2.3.2

pytest>=8.4.1
pytest-asyncio>=0.23.0

# Appenditions for primer subsystem. Keep existing lines; add these if missing.
biopython>=1.83
python-Levenshtein>=0.25

alembic>=1.13
